Move tx locking logic to a separate file
