+++ sui_core/src/lib.rs
@@ -11,3 +11,4 @@ pub mod consensus_client;
 pub mod execution_engine;
 pub mod gateway_state;
 pub mod safe_client;
+pub mod transaction_input_checker;
