+++ sui_core/src/unit_tests/authority_tests.rs
@@ -17,7 +17,7 @@ use sui_types::{
     crypto::{get_key_pair, Signature},
     gas::{calculate_module_publish_cost, get_gas_balance},
     messages::Transaction,
-    object::{GAS_VALUE_FOR_TESTING, OBJECT_START_VERSION},
+    object::{Owner, GAS_VALUE_FOR_TESTING, OBJECT_START_VERSION},
 };
 
 use std::fs;
