+++ sui_core/src/unit_tests/batch_transaction_tests.rs
@@ -13,6 +13,7 @@ use sui_adapter::genesis;
 use sui_types::{
     crypto::{get_key_pair, Signature},
     messages::Transaction,
+    object::Owner,
 };
 
 #[tokio::test]
